Add a generic disassembly pane over the system seam
Render disassembly at the Game Boy Instructions standard from the seam
alone: the walk and decode come from the core's instruction set, memory
from side-effect-free peeks or the per-vblank snapshot window, and every
row draws through a shared widget module both panes use, so the generic
Disassembly and the hand-built Instructions stay pixel-identical.

Seam additions carry the running-view fuel: snapshots gain the program
counter, symbols, code/data log, bank mapping, and instruction set;
SystemDebugger gains bank_for; InstructionSet gains classify_operand,
overridden by SM83 and the 6502. The stepping helper captures a
PC-anchored memory window into its snapshots via a new peek hook, so NES
and SMS get running disassembly (SMS falls back to raw byte rows). VCS
snapshots capture a window too.

Breakpoint message payloads widen to u32 so the pane's gutter works over
any bus width; cores mask to their own.

+++ crates/missingno/src/app/emu_thread.rs
@@ -72,8 +72,8 @@ pub enum EmuCommand {
     /// can hold the process open until teardown completes.
     Shutdown,
     SetControl(ControlId, ControlInput),
-    SetBreakpoint(u16),
-    ClearBreakpoint(u16),
+    SetBreakpoint(u32),
+    ClearBreakpoint(u32),
     AddWatchpoint(Watch),
     RemoveWatchpoint(Watch),
     RequestScreenshot {
@@ -467,17 +467,17 @@ impl Payload {
         }
     }
 
-    fn set_breakpoint(&mut self, address: u16) {
+    fn set_breakpoint(&mut self, address: u32) {
         match self {
             Self::Console(_) => {}
-            Self::Debugger(payload) => payload.core.set_breakpoint(address as u32),
+            Self::Debugger(payload) => payload.core.set_breakpoint(address),
         }
     }
 
-    fn clear_breakpoint(&mut self, address: u16) {
+    fn clear_breakpoint(&mut self, address: u32) {
         match self {
             Self::Console(_) => {}
-            Self::Debugger(payload) => payload.core.clear_breakpoint(address as u32),
+            Self::Debugger(payload) => payload.core.clear_breakpoint(address),
         }
     }
